requirements.txt
pony
urllib3
urllib3[socks]
python-telegram-bot
python-telegram-bot[socks]
